Moved executable utilities into their own package
