+++ strongback-executable/src/org/strongback/executable/utils/PropertiesUtils.java
@@ -1,4 +1,4 @@
-package org.strongback.executable;
+package org.strongback.executable.utils;
 
 import java.io.File;
 import java.io.FileNotFoundException;
